+++ hooks/post-tool-use.js
@@ -2,6 +2,7 @@
 'use strict';
 
 const { setDiscarded, isDiscardSignal } = require('./lib/discardGate.js');
+const { isMutatingNotebookTool, NUDGE_TEXT } = require('./lib/nudge.js');
 
 function main() {
   let event = {};
@@ -13,6 +14,10 @@ function main() {
   if (isDiscardSignal(event.tool_response)) {
     setDiscarded();
     process.stdout.write(JSON.stringify({ continue: false, stopReason: 'edit discarded' }) + '\n');
+  } else if (isMutatingNotebookTool(event.tool_name)) {
+    process.stdout.write(JSON.stringify({
+      hookSpecificOutput: { hookEventName: 'PostToolUse', additionalContext: NUDGE_TEXT },
+    }) + '\n');
   }
   process.exit(0);
 }